chore(pr) Moved  around to solve a reference issue.
