+++ editor/src/components/navigator/navigator-item/navigator-item.tsx
@@ -40,13 +40,25 @@ import {
   matchesOverriddenConditionalBranch,
 } from '../../../core/model/conditionals'
 import { DerivedSubstate, MetadataSubstate } from '../../editor/store/store-hook-substore-types'
-import {
-  getConditionalClausePathForNavigatorEntry,
-  getItemHeight,
-  navigatorDepth,
-} from '../navigator-utils'
+import { getConditionalClausePathForNavigatorEntry, navigatorDepth } from '../navigator-utils'
 import createCachedSelector from 're-reselect'
 import { getValueFromComplexMap } from '../../../utils/map'
+
+export function getItemHeight(navigatorEntry: NavigatorEntry): number {
+  if (isConditionalClauseNavigatorEntry(navigatorEntry)) {
+    if (navigatorEntry.clause === 'true-case') {
+      // The TRUE label row will be the shortest size.
+      return UtopiaTheme.layout.rowHeight.smallest
+    } else {
+      // The FALSE label row should visually appear to be the shortest size.
+      // The size difference (against the TRUE label row) will be a top margin.
+      return UtopiaTheme.layout.rowHeight.smaller
+    }
+  } else {
+    // Default size for everything else.
+    return UtopiaTheme.layout.rowHeight.smaller
+  }
+}
 
 export const NavigatorItemTestId = (pathString: string): string =>
   `NavigatorItemTestId-${pathString}`
